Refactor rxing-wasm integration
Avoid rescanning the entire page with rxing-wasm, to speed up the
scanning process. Rewrite the rxing scan function to accept the
page canvas context and the current barcode's bounding box to
extract the image data and only rescan that area.

A white 5px padding border is added to the temporary image data,
to increase the success rate of the rxing rescan.

Signed-off-by: Stefan Knoblich <stkn@bitplumber.de>